{{cookiecutter.project.name}}/requirements.txt
debugpy
django
djangorestframework==3.13.1
markdown
django-filter
django-extensions
Werkzeug
django-allauth
django-cors-headers
social-auth-app-django
social-auth-core[openidconnect]
rest-social-auth
dj-rest-auth

django-storages
drf-access-policy
drf-nested-routers
drf-yasg
pylint-django
django-redis
whitenoise


django-rest-framework-social-oauth2
django-request-logging
django-health-check
django-oauth-toolkit
pytest-django
pytest-factoryboy
pytest-cov
python-decouple

whitenoise
gunicorn
django-storages
psycopg2
mysqlclient
boto3

{% if cookiecutter.configuration.xray.enabled == 'True' and cookiecutter.deployment.selected != 'EYK' %}
aws-xray-sdk
tox
{% endif %}
